pyproject.toml
[project]
name = "seinfeld-script-generator"
version = "0.1.0"
description = "A multi-agent system using CrewAI to generate Seinfeld episode scripts using Couchbase Capella AI Services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]>=0.86.0",
    "litellm>=1.0.0",
    "langchain-openai==0.3.5",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langchain-couchbase==0.5.0",
]

[project.scripts]
seinfeld = "seinfeld_script_generator.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/seinfeld_script_generator"]

